Fix ath6003 firmware for sequence error on 802.11n
The firmware would stop sending data back up to the driver after about 10
seconds when connected to an AP in N rates. Turns out to have been an
error with the sequence numbers. This is an update to the firmware to
version 3.4.0.80 that fixes the bug and reinstates all previous functionality
that had been rolled back when we backed the firmware back to a working
version.

The following features are reinstated or fixed by this firmware:

*  Added API for ACK/CTS timeout
*  Adds "peek" API for reading hardware registers
*  Increased returned channel list size from 32 to 64. The host driver
     must also change the "WMI_MAX_CHANNELS" macro in "wmi.h" from 32
     to 64 for compatibility.
*  Fixes the ad-hoc communication problem (ticket #5683838)
*  Contains the fix for the sdc_cli Device Name not working (ticket #6167489)
*  Fixed ticket 6438323, which caused the WB45 to stop passing traffic
     when attached to an 802.11n AP.

Bug: 4723