Improve /e/n/i file handling with atomic and backup-copy methods.
Added: ifrc uses file copy/move method for changing "auto <iface>".
...the temporary copy has an appended tilde character.

Added: Both the nis and ifrc will check for not-zero-length interfaces.
...corrective action is to check for not-zero-length interfaces~ file,
and if okay, then will use it as the backup working copy.
(And as a last resort, use the gzip'd backup copy.)

Added to post-build scripts: Command automatically creates a gzip'd copy
of the /e/n/i file as "interfaces~.gz".  The nis and ifrc will then use
a decompressed copy of this file, as last resort.

BZ_4846