Allow other projects to reference sdk_version.h
Similar scope to config_strings.h. We don’t make it public (to customers)
but we make it available for other projects (supplicant) to reference.
The goal is to have any project that references the sdk to get its first
two version digits directly from sdk_version.h, so it is less trouble when
 we bump versions of the sdk.