Add prune-workspace target
This new target will allow Jenkins or other CI systems to hugely reduce
the disk-space footprint. It should only be utilized after a successful
build on a workspace you don't intend to build in again, because it will
totally crush the workspace. Thus, it's not intended for developers to
use.

Bug: 8657