laird: Allow customer modified fw.txt to be used
The existing system generates two fw.txt, one from the post-image-common
script called from post-image, which is discarded.
And another from the a common.mk build step after this point.

Remove the re-generation and copy the version that can be fixed up by
the customer post-image.sh to allow for additional locations to be copied.

Bug 10502