sterling_packages.mk: start tarball at lib and remove preserve attributes
In order to avoid changing users and permissions on the current directory
the tarball is extracted in, start taring at "lib" instead of ".".
Also, remove preserve all attributes on the copy.  Git stores all files
with a default set of permissions, ignores user, and only cares about the
execute bit.  There isn't a great reason to preserve file permissions or
owners.

Bug: 11263