sterling-lwb-firmware: selectively add .clm_blob file
Currently the 4343w firmware package contains a .clm_blob file while the
4339 does not, but is expected to have one in the future.  Here the
determination of including the .clm_blob file is done by checking if one
exists in the main repo, and if so, adding it as a dependency.
The variable will be empty if the file does not exist so no dependency
will exist.